--- python/ray/rllib/io/__init__.py
@@ -1,16 +1,0 @@
-from ray.rllib.io.io_context import IOContext
-from ray.rllib.io.json_reader import JsonReader
-from ray.rllib.io.json_writer import JsonWriter
-from ray.rllib.io.output_writer import OutputWriter, NoopOutput
-from ray.rllib.io.input_reader import InputReader
-from ray.rllib.io.mixed_input import MixedInput
-
-__all__ = [
-    "IOContext",
-    "JsonReader",
-    "JsonWriter",
-    "NoopOutput",
-    "OutputWriter",
-    "InputReader",
-    "MixedInput",
-]
