+++ python/ray/rllib/offline/json_reader.py
@@ -14,7 +14,7 @@ try:
 except ImportError:
     smart_open = None
 
-from ray.rllib.io.input_reader import InputReader
+from ray.rllib.offline.input_reader import InputReader
 from ray.rllib.evaluation.sample_batch import SampleBatch
 from ray.rllib.utils.annotations import override
 from ray.rllib.utils.compression import unpack_if_needed
